+++ packages/__tests__/src/icalendar/day-view.ts
@@ -38,6 +38,12 @@ describe('addICalEventSource with day view', () => {
     loadICalendarWith(recurringWeekly, () => {
       setTimeout(() => {
         assertEventCount(1)
+        const event = currentCalendar.getEvents()[0]
+        // test non-date props
+        expect(event.title).toBe('Weekly Monday meeting')
+        expect(event.url).toBe('https://fullcalendar.io/')
+        expect(event.extendedProps.description).toBe('this is the description')
+        expect(event.extendedProps.location).toBe('this is the location')
         done()
       }, 100)
     })
@@ -47,7 +53,13 @@ describe('addICalEventSource with day view', () => {
     loadICalendarWith(alldayEvent, () => {
       setTimeout(() => {
         assertEventCount(1)
-        currentCalendar.getEvents().forEach((event) => expect(event.allDay).toBeTruthy())
+        const events = currentCalendar.getEvents()
+        events.forEach((event) => expect(event.allDay).toBeTruthy())
+        // test non-date props
+        expect(events[0].title).toBe('First conference')
+        expect(events[0].url).toBe('https://fullcalendar.io/')
+        expect(events[0].extendedProps.description).toBe('this is the description')
+        expect(events[0].extendedProps.location).toBe('this is the location')
         done()
       })
     })